git-svn-id: svn://svn.code.sf.net/p/x10/code/branches/dsl@29712 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/x10rt/mpi/Java.cc
@@ -33,7 +33,7 @@ int main(int ac, char* av[]) {
     int vmNOptions = clsIndex - 1;
     char* clsName = av[clsIndex];
     char* clsNameFQ = strdup(clsName);
-    for (int i = 0; i < strlen(clsNameFQ); ++i) {
+    for (size_t i = 0; i < strlen(clsNameFQ); ++i) {
         if (clsNameFQ[i] == '.') clsNameFQ[i] = '/';
     }
     char** clsOptions = av + clsIndex + 1;
